% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-package.R
\docType{package}
\name{consumer.pkg-package}
\alias{consumer.pkg}
\alias{consumer.pkg-package}
\title{consumer.pkg: Cross-Package Trait Dispatch Consumer Example}
\description{
Demonstrates miniextendr's cross-package trait dispatch capabilities. Provides generic functions that operate on the Counter trait interface. Works with Counter implementations from any package (like producer.pkg), proving ABI compatibility via type tag matching and vtable dispatch.
}
\author{
\strong{Maintainer}: Mossa Merhi Reimert \email{mossa@struchstrup.com}

Other contributors:
\itemize{
  \item Eleanor Andersson [contributor]
}

}
\keyword{internal}
