% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{consumer_get_class}
\alias{consumer_get_class}
\title{consumer_get_class}
\usage{
consumer_get_class(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
Character vector of class names
}
\description{
Get class of any R object (for cross-package testing)
}
