% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{consumer_greet}
\alias{consumer_greet}
\title{consumer_greet}
\usage{
consumer_greet(name)
}
\arguments{
\item{name}{Name to greet}
}
\value{
A greeting string
}
\description{
Greet with a message demonstrating consumer package is working
}
