% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{consumer_magic_number}
\alias{consumer_magic_number}
\title{consumer_magic_number}
\usage{
consumer_magic_number()
}
\value{
The number 42
}
\description{
Return a constant to verify the package is loaded
}
