% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{debug_consumer_tag_resettable}
\alias{debug_consumer_tag_resettable}
\title{debug_consumer_tag_resettable}
\usage{
debug_consumer_tag_resettable()
}
\description{
Debug: Get TAG_RESETTABLE as hex string
}
