% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{has_class}
\alias{has_class}
\title{has_class}
\usage{
has_class(x, class_name)
}
\arguments{
\item{x}{Any R object}

\item{class_name}{Class name to check for}
}
\value{
TRUE if object has the specified class
}
\description{
Check if an object has a specific class
}
