% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{is_counter}
\alias{is_counter}
\title{is_counter}
\usage{
is_counter(sexp)
}
\arguments{
\item{sexp}{Any R object}
}
\value{
TRUE if the object implements Counter trait
}
\description{
Check if an object implements the Counter trait
}
