% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{is_external_ptr}
\alias{is_external_ptr}
\title{is_external_ptr}
\usage{
is_external_ptr(sexp)
}
\arguments{
\item{sexp}{Any R object}
}
\value{
TRUE if it's an ExternalPtr
}
\description{
Check if a SEXP is an ExternalPtr (type-agnostic check)
}
