% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer-pkg-wrappers.R
\name{add_and_get}
\alias{add_and_get}
\title{Add a value to a counter and return the new value}
\source{
Generated by miniextendr from Rust fn \code{add_and_get}
}
\usage{
add_and_get(counter_sexp, n)
}
\arguments{
\item{counter_sexp}{An ExternalPtr to any type implementing Counter}

\item{n}{Value to add}
}
\value{
The counter's value after adding n
}
\description{
Add a value to a counter and return the new value
}
