% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{check_is_default}
\alias{check_is_default}
\title{Check if an object is in its default state (without resetting)}
\source{
Generated by miniextendr from Rust fn \code{check_is_default}
}
\usage{
check_is_default(sexp)
}
\arguments{
\item{sexp}{An ExternalPtr to any type implementing Resettable}
}
\value{
TRUE if the object is in default state
}
\description{
Check if an object is in its default state (without resetting)
}
