% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{DoubleCounter}
\alias{DoubleCounter}
\alias{DoubleCounter$get_value}
\alias{DoubleCounter$create}
\alias{$.DoubleCounter}
\alias{[[.DoubleCounter}
\alias{DoubleCounter$Counter$value}
\alias{DoubleCounter$Counter$increment}
\alias{DoubleCounter$Counter$add}
\alias{DoubleCounter$Counter$panic_plain}
\alias{DoubleCounter$Counter$error_with_class}
\alias{DoubleCounter$Counter$raise_error}
\alias{DoubleCounter$Counter$raise_warning}
\alias{DoubleCounter$Counter$raise_message}
\alias{DoubleCounter$Counter$raise_condition_classed}
\alias{DoubleCounter$Counter$raise_error_with_data}
\title{DoubleCounter  Class}
\usage{
DoubleCounter

\method{$}{DoubleCounter}(self, name)

\method{[[}{DoubleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get the current value (inherent method)
}
