% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer-pkg-wrappers.R
\docType{data}
\name{DoubleCounter}
\alias{DoubleCounter}
\alias{DoubleCounter$get_value}
\alias{DoubleCounter$create}
\alias{$.DoubleCounter}
\alias{[[.DoubleCounter}
\alias{DoubleCounter$Counter$value}
\alias{DoubleCounter$Counter$increment}
\alias{DoubleCounter$Counter$add}
\title{DoubleCounter  Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{DoubleCounter}

Generated by miniextendr from \code{DoubleCounter::get_value}

Generated by miniextendr from \code{DoubleCounter::create}
}
\usage{
DoubleCounter

\method{$}{DoubleCounter}(self, name)

\method{[[}{DoubleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get the current value (inherent method)
}
\keyword{datasets}
