% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{DoubleCounter}
\alias{DoubleCounter}
\alias{DoubleCounter$get_value}
\alias{DoubleCounter$create}
\alias{$.DoubleCounter}
\alias{[[.DoubleCounter}
\alias{DoubleCounter$Counter$value}
\alias{DoubleCounter$Counter$increment}
\alias{DoubleCounter$Counter$add}
\title{DoubleCounter Class}
\format{
An object of class \code{environment} of length 3.
}
\source{
Generated by miniextendr from Rust type \code{DoubleCounter}
}
\usage{
DoubleCounter

\method{$}{DoubleCounter}(self, name)

\method{[[}{DoubleCounter}(self, name)
}
\description{
Get the current value (inherent method)
}
\keyword{datasets}
