% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer-pkg-wrappers.R
\name{get_reset_get}
\alias{get_reset_get}
\title{Get the value of a Counter, reset it, then get the value again Returns the value after reset}
\source{
Generated by miniextendr from Rust fn \code{get_reset_get}
}
\usage{
get_reset_get(sexp)
}
\arguments{
\item{sexp}{An ExternalPtr to a type implementing BOTH Counter and Resettable}
}
\value{
The counter value after reset
}
\description{
Get the value of a Counter, reset it, then get the value again Returns the value after reset
}
