% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{increment_twice}
\alias{increment_twice}
\title{increment_twice}
\usage{
increment_twice(counter_sexp)
}
\arguments{
\item{counter_sexp}{An ExternalPtr to any type implementing Counter}
}
\value{
The counter's value after incrementing twice
}
\description{
Increment a counter twice (generic over Counter trait) This function works with Counter objects from ANY package.
}
