% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{is_external_ptr}
\alias{is_external_ptr}
\title{Check if a SEXP is an ExternalPtr (type-agnostic check)}
\source{
Generated by miniextendr from Rust fn \code{is_external_ptr}
}
\usage{
is_external_ptr(sexp)
}
\arguments{
\item{sexp}{Any R object}
}
\value{
TRUE if it's an ExternalPtr
}
\description{
Check if a SEXP is an ExternalPtr (type-agnostic check)
}
