% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{is_resettable}
\alias{is_resettable}
\title{Check if an object implements the Resettable trait}
\source{
Generated by miniextendr from Rust fn \code{is_resettable}
}
\usage{
is_resettable(sexp)
}
\arguments{
\item{sexp}{Any R object}
}
\value{
TRUE if the object implements Resettable trait
}
\description{
Check if an object implements the Resettable trait
}
