% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer-pkg-wrappers.R
\name{passthrough_ptr}
\alias{passthrough_ptr}
\title{Pass an ExternalPtr through consumer without knowing its type}
\source{
Generated by miniextendr from Rust fn \code{passthrough_ptr}
}
\usage{
passthrough_ptr(ptr)
}
\arguments{
\item{ptr}{An ExternalPtr from any package}
}
\value{
The same ExternalPtr (pass-through)
}
\description{
Pass an ExternalPtr through consumer without knowing its type
}
