% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.pkg-wrappers.R
\name{reset_and_check}
\alias{reset_and_check}
\title{Reset an object and check if it's in default state This function works with Resettable objects from ANY package}
\source{
Generated by miniextendr from Rust fn \code{reset_and_check}
}
\usage{
reset_and_check(sexp)
}
\arguments{
\item{sexp}{An ExternalPtr to any type implementing Resettable}
}
\value{
TRUE if the object is in default state after reset
}
\description{
Reset an object and check if it's in default state This function works with Resettable objects from ANY package
}
